cleaning up signin-form and improving UI
