+++ pages/login.html
@@ -13,7 +13,13 @@
       href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&family=Open+Sans:ital,wght@0,300..800;1,300..800&display=swap"
       rel="stylesheet"
     />
-    <script src="../src/js/index.js" type="module" defer></script>
+    <script type="module" src="../src/js/index.js" defer></script>
+    <script
+      type="module"
+      src="../src/js/pagesJS/loginSignup.mjs"
+      defer
+    ></script>
+    c
     <link rel="stylesheet" href="../dist/css/index.css" />
     <title>Welcome | Noroff Auction House</title>
   </head>
