+++ src/js/api-functions/post/postActions.mjs
@@ -1,4 +1,5 @@
 import { postData } from "./postData.mjs";
+import { changeForm } from "../../visual-effects/loginEffects.mjs";
 
 function goToPage(page) {
   window.location.href = [`${page}`];
@@ -10,20 +11,37 @@ function goToPage(page) {
  * @param {object} formData - what data is sent via the request body
  * @param {object} divForError - a div where error messages are printed
  *
- * This function logs you into the social media platform
- *
- * NOTE TO ME; Needs to work without switch
+ * This function logs you in
  */
-export function login(url, formData, divForError, formName) {
+export function login(url, formData, divForError) {
   const headers = {
     "Content-Type": "application/json",
   };
-  //   let page;
-  console.log(url);
-  console.log(formData);
   const page = "/index.html";
-
   postData(url, formData, headers, divForError, goToPage, page);
+}
+
+/**
+ *
+ * @param {string} url - where to send the data
+ * @param {object} formData - what data is sent via the request body
+ * @param {object} divForError - a div where error messages are printed
+ * @param {*} mainDOMElement - the moving main DOM-element
+ *
+ * registers a new user. if no error occurs, you are taken to the login-portal
+ * as the API does not return a token if you register, only if you log in
+ */
+export function signup(url, formData, divForError, mainDOMElement) {
+  const headers = {
+    "Content-Type": "application/json",
+  };
+  divForError.innerHTML = "";
+  postData(url, formData, headers, divForError);
+  setTimeout(() => {
+    if (divForError.innerHTML === "") {
+      changeForm(mainDOMElement);
+    }
+  }, 1000);
 }
 
 /**
